abstracted municip map to LocatorMap
