+++ scripts/components/factsheets/area-stack.component.js
