+++ scripts/components/factsheets/chord.component.js
