+++ scripts/components/factsheets/table.component.js
